Undefined behaviour is the spice of life

+++ src/vm/load_store.c
@@ -6,7 +6,7 @@
 /*   By: aamadori <aamadori@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/15 15:19:53 by aamadori          #+#    #+#             */
-/*   Updated: 2019/03/20 14:45:33 by aamadori         ###   ########.fr       */
+/*   Updated: 2019/03/20 18:03:07 by aamadori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 /* TODO always load and store memory through these functions */
 /* TODO compile-time ASSERT if REG_SIZE is too big */
-t_bigend_buffer	mem_load(t_vm_state *state, size_t address, size_t size)
+t_bigend_buffer	mem_load(t_vm_state *state, int64_t address, size_t size)
 {
 	t_bigend_buffer	ret;
 	size_t			byte_address;
@@ -40,7 +40,7 @@ t_bigend_buffer	mem_load(t_vm_state *state, size_t address, size_t size)
 	return	ret;
 }
 
-void	mem_store(t_vm_state *state, size_t address, size_t size,
+void	mem_store(t_vm_state *state, int64_t address, size_t size,
 			const t_bigend_buffer store)
 {
 	size_t			byte_address;
